+++ bgameb/base.py
@@ -1,9 +1,10 @@
 """Base constructs for build package objects
 """
-from typing import Dict, List, Optional, Any, Iterator
+from typing import Dict, List, Optional, Any, Iterator, Tuple
 from collections.abc import Mapping
 from dataclasses import dataclass, field, make_dataclass
-from dataclasses_json import DataClassJsonMixin, config
+from heapq import heappop, heappush
+from dataclasses_json import config, dataclass_json
 from bgameb.errors import ComponentNameError
 from loguru import logger
 
@@ -32,8 +33,9 @@ def log_enable(
     logger.enable('bgameb')
 
 
-@dataclass(init=False)
-class Components(Mapping, DataClassJsonMixin):
+@dataclass_json
+@dataclass(repr=False)
+class Components(Mapping):
     """Components mapping
     """
     def __init__(
@@ -77,7 +79,11 @@ class Components(Mapping, DataClassJsonMixin):
         del self.__dict__[attr]
 
     def __repr__(self):
-        items = (f"{k}={v!r}" for k, v in self.__dict__.items())
+        items = (
+            f"{k}={v!r}" for k, v
+            in self.__dict__.items()
+            if not k.startswith('_') and not k.startswith('current')
+            )
         return "{}({})".format(type(self).__name__, ", ".join(items))
 
     def __len__(self) -> int:
@@ -87,13 +93,13 @@ class Components(Mapping, DataClassJsonMixin):
         """Chek is name of component is unique
 
         Args:
-            name (str): name for component
+            name (str): name of component
 
         Raises:
             ComponentNameError: name id not unique
 
         Returns:
-            Optional[bool]: if it is in
+            Optional[bool]: if it is in Components and unique
         """
         if name in self.__dict__.keys():
             raise ComponentNameError(name=name)
@@ -108,7 +114,7 @@ class Components(Mapping, DataClassJsonMixin):
 
         Args:
             component: component class
-            kwargs (Dict[str, Any]): aditional args
+            kwargs (Dict[str, Any]): additional args for component
         """
         comp = component(**kwargs)
 
@@ -117,20 +123,21 @@ class Components(Mapping, DataClassJsonMixin):
                 self.__class__.__name__,
                 fields=[(kwargs['name'], type(comp), field(default=comp))],
                 bases=(self.__class__, ),
+                repr=False
                 )
 
         self.__dict__.update({kwargs['name']: comp})
 
     def _add(self, component, **kwargs) -> None:
         """Add component to Components dict. Components with
-        same names as existed cant be added.
+        same names as existed in Components cant be added.
 
         Raises:
-            ComponentNameError: name id not unique
+            ComponentNameError: name not unique
 
         Args:
-            component (Component): component class
-            kwargs: aditional args
+            component: component class
+            kwargs: additional args for component
         """
         if kwargs.get('name'):
             self._chek_in(kwargs['name'])
@@ -144,28 +151,78 @@ class Components(Mapping, DataClassJsonMixin):
         """Add or replace component in Components dict.
 
         Args:
-            component (Component): component class
-            kwargs: aditional args
+            component: component class
+            kwargs: additional args for component
         """
         if not kwargs.get('name'):
             kwargs['name'] = component.name
         self._update(component, kwargs)
 
     def get_names(self) -> List[str]:
-        """Get names of all components of class
+        """Get names of all components in Components
 
         Returns:
-            List[str]: lisct of names of conatined components
+            List[str]: list of components names
         """
         return list(self.__dict__)
 
 
+@dataclass_json
 @dataclass
-class Base(Components, DataClassJsonMixin):
+class Order:
+    """Order of steps priority queue. Isnt tradesafe.
+    Can be used for define geme steps order.
+
+    Args:
+
+        - current List[Tuple[int, Components]]: priority queue list
+    """
+    current: List[Tuple[int, Components]] = field(
+        default_factory=list,
+        metadata=config(exclude=lambda x: True),  # type: ignore
+        repr=False,
+            )
+
+    def __len__(self) -> int:
+        """Len of queue
+
+        Returns:
+            int: len of current queue
+        """
+        return len(self.current)
+
+    def clear(self) -> None:
+        """Clear the current queue
+        """
+        self.current = []
+
+    def put(self, item) -> None:
+        """Put Step object to queue
+
+        Args:
+            item (Step): Step class instance
+        """
+        heappush(self.current, (item.priority, item))
+
+    def get(self) -> Components:
+        """Get Syep object from queue with lowest priority
+
+        Returns:
+            Step: Step instance object
+        """
+        return heappop(self.current)[1]
+
+
+@dataclass_json
+@dataclass(repr=False)
+class Base(Components):
     """Base class for game, stuff, tools players and other components
+
+    Args:
+
+        - name (str): name of component
     """
     name: str
-    is_active: bool = True
 
     def __post_init__(self) -> None:
         # check name
@@ -175,9 +232,9 @@ class Base(Components, DataClassJsonMixin):
             )
 
         # set logger
-        self.logger = logger.bind(
+        self._logger = logger.bind(
             classname=self.__class__.__name__,
             name=self.name)
-        self.logger.info(
+        self._logger.info(
             f'{self.__class__.__name__} created with {self.name=}.'
             )