+++ bgameb/players.py
@@ -3,23 +3,23 @@
 from typing import Optional, Literal, List
 from collections import Counter
 from dataclasses import dataclass, field
-from dataclasses_json import config
+from dataclasses_json import config, dataclass_json
 from bgameb.base import Base
 
 
-@dataclass
+@dataclass_json
+@dataclass(repr=False)
 class BasePlayer(Base):
     """Base class for game players and bots
 
-    Attrs
+    Args:
 
         - name (str): player name
         - counter (Counter): counter object for count any items
-        - is_active (bool): Default to True.
-        - has_priority (bool): Default to False
-        - team (str, optioanl): team name for this player
-        - owner_of (List[str]): list of objects owned by player Default to []
-        - user_of (List[str]): list of objects used by player Default to []
+        - has_priority (bool): Priority flag. Default to False
+        - team (str, optioanl): team name for player. Default to None
+        - owner_of (List[str]): list of object names owned by
+                                player Default to []
 
     """
     counter: Counter = field(
@@ -30,18 +30,17 @@ class BasePlayer(Base):
     has_priority: bool = False
     team: Optional[str] = None
     owner_of: List[str] = field(default_factory=list)
-    user_of: List[str] = field(default_factory=list)
-    rules: List[str] = field(default_factory=list)
 
     def __post_init__(self) -> None:
         super().__post_init__()
 
 
-@dataclass
+@dataclass_json
+@dataclass(repr=False)
 class Player(BasePlayer):
-    """Base class to create a player
+    """Player class
     """
-    type_: str = field(
+    _type: str = field(
         default='player',
         metadata=config(exclude=lambda x: True),  # type: ignore
         repr=False
@@ -51,7 +50,5 @@ class Player(BasePlayer):
         super().__post_init__()
 
 
-PLAYERS = {
-    Player.type_: Player,
-    }
+PLAYERS = {Player._type: Player, }
 PLAERS_TYPES = Literal['player', ]